git add src/controllers/teamApplicationsController.js src/controllers/teamController.js src/routes/teamRoutes.js test/teamController.applications.test.js test/teamController.applyToJoinTeam.test.js
